+++ src/core_modules/capture-core/components/WidgetEventSchedule/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { WidgetEventSchedule } from './WidgetEventSchedule.container';