+++ src/core_modules/capture-core/converters/clientToList.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import { moment } from 'capture-core-utils/moment';
+import moment from 'moment';
 import { Tag } from '@dhis2/ui';
 import { dataElementTypes, type DataElement } from '../metaData';
 import { convertMomentToDateFormatString } from '../utils/converters/date';