+++ src/core_modules/capture-core/utils/routing/index.js
@@ -1,3 +1,4 @@
 // @flow
 export { useLocationQuery } from './useLocationQuery';
 export { deriveURLParamsFromHistory } from './deriveURLParamsFromHistory';
+export { buildUrlQueryString } from './buildUrlQueryString';